Add two new folder and moved the previous wdf folders
