+++ src/app/features/wdf/wdf-data-change/wdf-staff-record-status-message/wdf-staff-record-status-message.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 
-import { Worker } from '../../../core/model/worker.model';
+import { Worker } from '../../../../core/model/worker.model';
 
 @Component({
   selector: 'app-wdf-staff-record-status-message',
