+++ src/app/features/wdf/wdf-data-change/wdf-staff-record/wdf-pagination/wdf-pagination.component.spec.ts
@@ -3,7 +3,7 @@ import { getTestBed } from '@angular/core/testing';
 import { BrowserModule, By } from '@angular/platform-browser';
 import { ActivatedRoute, NavigationEnd, Router, RouterEvent } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
-import { WdfPaginationComponent } from '@features/wdf/wdf-staff-record/wdf-pagination/wdf-pagination.component';
+import { WdfPaginationComponent } from '@features/wdf/wdf-data-change/wdf-staff-record/wdf-pagination/wdf-pagination.component';
 import { SharedModule } from '@shared/shared.module';
 import { render } from '@testing-library/angular';
 import { Observable, Subject } from 'rxjs';
