+++ src/app/features/wdf/wdf-data-change/wdf-staff-record/wdf-pagination/wdf-pagination.component.ts
@@ -3,7 +3,7 @@ import { ActivatedRoute } from '@angular/router';
 
 @Component({
   selector: 'app-wdf-pagination',
-  templateUrl: './wdf-pagination.component.html'
+  templateUrl: './wdf-pagination.component.html',
 })
 export class WdfPaginationComponent implements OnInit {
   @Input() workerList: string[];
@@ -11,21 +11,19 @@ export class WdfPaginationComponent implements OnInit {
   public isLast = false;
   public nextID: string;
   public previousID: string;
-  constructor(
-    private route: ActivatedRoute,
-  ) { }
+  constructor(private route: ActivatedRoute) {}
 
   ngOnInit(): void {
     this.route.params.subscribe((data) => {
-       this.setVariables(data)
-    })
+      this.setVariables(data);
+    });
   }
-  public setVariables(data){
+  public setVariables(data) {
     const workerUID = data.id;
-    const workerIndex = this.workerList.findIndex((uid) =>  uid === workerUID);
+    const workerIndex = this.workerList.findIndex((uid) => uid === workerUID);
     this.isFirst = workerIndex === 0;
     this.isLast = workerIndex === this.workerList.length - 1;
-    this.nextID = this.workerList[workerIndex+1];
-    this.previousID = this.workerList[workerIndex-1];
+    this.nextID = this.workerList[workerIndex + 1];
+    this.previousID = this.workerList[workerIndex - 1];
   }
 }
