+++ src/app/features/wdf/wdf-data-change/wdf-workplaces-summary-table/wdf-workplaces-summary-table.component.ts
