+++ src/app/features/wdf/wdf-data-change/wdf-workplaces-summary/wdf-workplaces-summary.component.html
