+++ src/app/features/wdf/wdf-data-change/wdf.module.ts
@@ -10,9 +10,7 @@ import { WdfOverviewComponent } from './wdf-overview/wdf-overview.component';
 import { WdfParentStatusMessageComponent } from './wdf-parent-status-message/wdf-parent-status-message.component';
 import { WdfRequirementsStateComponent } from './wdf-requirements-state/wdf-requirements-state.component';
 import { WdfRoutingModule } from './wdf-routing.module';
-import {
-  WdfStaffRecordStatusMessageComponent,
-} from './wdf-staff-record-status-message/wdf-staff-record-status-message.component';
+import { WdfStaffRecordStatusMessageComponent } from './wdf-staff-record-status-message/wdf-staff-record-status-message.component';
 import { WdfPaginationComponent } from './wdf-staff-record/wdf-pagination/wdf-pagination.component';
 import { WdfStaffRecordComponent } from './wdf-staff-record/wdf-staff-record.component';
 import { WdfStaffSummaryComponent } from './wdf-staff-summary/wdf-staff-summary.component';
