--- src/app/features/wdf/wdf-workplaces-summary-table/wdf-workplaces-summary-table.component.html
@@ -1,48 +1,0 @@
-<div class="govuk-form-group">
-  <label class="govuk-label" for="sortByStaff"> Sort by </label>
-  <select class="govuk-select" id="sortByStaff" name="sortByStaff" (change)="sortByColumn($event.target.value)">
-    <option
-      *ngFor="let sortWorkplaceOption of sortWorkplacesOptions | keyvalue"
-      value="{{ sortWorkplaceOption.key }}"
-      [selected]="sortWorkplaceOption.key == this.sortBy"
-    >
-      {{ sortWorkplaceOption.value }}
-    </option>
-  </select>
-</div>
-
-<table *ngIf="workplaces.length" class="govuk-table govuk-table">
-  <thead class="govuk-table__head">
-  <tr class="govuk-table__row">
-    <th scope="col" class="govuk-table__header">Workplace name</th>
-    <th scope="col" class="govuk-table__header">WDF requirements (workplace)</th>
-    <th scope="col" class="govuk-table__header">WDF requirements (staff records)</th>
-  </tr>
-  </thead>
-  <tbody class="govuk-table__body">
-  <tr *ngFor="let workplace of workplaces" class="govuk-table__row">
-    <td class="govuk-table__cell">
-      <ng-container *ngIf="canViewWorkplace(workplace); else noLink">
-        <a [routerLink]="[workplace.uid]">
-          {{ workplace.name }}
-        </a>
-      </ng-container>
-      <ng-template #noLink>
-        <p> {{ workplace.name }}</p>
-      </ng-template>
-    </td>
-    <td class="govuk-table__cell">
-      <app-wdf-requirements-state
-        [overallWdfEligibility]="workplace.wdf.overall"
-        [currentWdfEligibility]="workplace.wdf.workplace"
-      ></app-wdf-requirements-state>
-    </td>
-    <td class="govuk-table__cell">
-      <app-wdf-requirements-state
-        [overallWdfEligibility]="workplace.wdf.overall"
-        [currentWdfEligibility]="workplace.wdf.staff"
-      ></app-wdf-requirements-state>
-    </td>
-  </tr>
-  </tbody>
-</table>
